Assignment-3/Assignment-3/HouseRentSystem/tests/test_app.py
import pytest
import os
import tempfile
import sqlite3
from flask import session
from io import BytesIO
from app import app, init_db, init_property_db, init_messages_db

@pytest.fixture
def client():
    db_fd, app.config['DATABASE'] = tempfile.mkstemp()
    app.config['TESTING'] = True

    with app.test_client() as client:
        with app.app_context():
            init_db()
            init_property_db()
            init_messages_db()
        yield client

    os.close(db_fd)
    os.unlink(app.config['DATABASE'])

def test_init_db(client):
    """Test to ensure database is created properly."""
    conn = sqlite3.connect('rental_management.db')
    cursor = conn.cursor()

    # Check if users table exists
    cursor.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='users'")
    users_table = cursor.fetchone()
    assert users_table is not None

    # Check if properties table exists
    cursor.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='properties'")
    properties_table = cursor.fetchone()
    assert properties_table is not None

    # Check if messages table exists
    cursor.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='messages'")
    messages_table = cursor.fetchone()
    assert messages_table is not None

    conn.close()

def test_register(client):
    """Test the registration endpoint."""
    response = client.post('/register', data={
        'phone': '1234567890',
        'username': 'testuser',
        'email': 'test@example.com',
        'password': 'password123',
        'user_type': 'tenant'
    })

    # Check if the status code is 200
    assert response.status_code == 200

    # Verify that the response contains the form HTML elements, indicating the request did not fail
    assert b'<form' in response.data and b'name="username"' in response.data


def test_login(client):
    """Test the login endpoint."""
    # Register user first
    client.post('/register', data={
        'phone': '1234567890',
        'username': 'testuser',
        'email': 'test@example.com',
        'password': 'password123',
        'user_type': 'tenant'
    })

    # Test login
    response = client.post('/login', data={
        'username': 'testuser',
        'password': 'password123'
    })
    assert response.status_code == 200
    assert b'success' in response.data


def test_add_property(client):
    """Test adding a property for a landlord."""
    # Register landlord
    client.post('/register', data={
        'phone': '1234567890',
        'username': 'landlorduser',
        'email': 'landlord@example.com',
        'password': 'password123',
        'user_type': 'landlord'
    })

    # Login and set session manually
    with client.session_transaction() as sess:
        sess['user_id'] = 1
        sess['username'] = 'landlorduser'
        sess['user_type'] = 'landlord'
        sess['phone'] = '1234567890'

    # Test adding a property, skipping specific field checks
    response = client.post('/add_property', data={
        'propertyType': 'Apartment',
        'propertyAddress': '123 Street Name',
        'propertyZipCode': '12345',
        'leaseTerm': '12 months',
        'rentRate': '1000',
        'availableDate': '2024-11-01'
    }, follow_redirects=True)

    # Check only if status code is 200 (indicating success after redirect)
    assert response.status_code == 200

def test_delete_property(client):
    """Test deleting a property."""
    # Register landlord and add property
    client.post('/register', data={
        'phone': '1234567890',
        'username': 'landlorduser',
        'email': 'landlord@example.com',
        'password': 'password123',
        'user_type': 'landlord'
    })

    client.post('/login', data={
        'username': 'landlorduser',
        'password': 'password123'
    })

    client.post('/add_property', data={
        'propertyType': 'Apartment',
        'propertyAddress': '123 Street Name',
        'propertyZipCode': '12345',
        'leaseTerm': '12 months',
        'rentRate': '1000',
        'availableDate': '2024-11-01'
    })

    # Get property ID
    conn = sqlite3.connect('rental_management.db')
    cursor = conn.cursor()
    cursor.execute('SELECT id FROM properties WHERE landlord_phone = ?', ('1234567890',))
    property_id = cursor.fetchone()[0]
    conn.close()

    # Set the phone in session to avoid BuildError
    with client.session_transaction() as sess:
        sess['phone'] = '1234567890'  # Simulate phone in session

    # Test deleting the property
    response = client.post(f'/delete_property/{property_id}', follow_redirects=False)

    # Check if response is a redirect (302 status code)
    assert response.status_code == 302


def test_logout(client):
    """Test logging out."""
    # Register and log in a user to simulate a session
    client.post('/register', data={
        'phone': '1234567890',
        'username': 'testuser',
        'email': 'test@example.com',
        'password': 'password123',
        'user_type': 'tenant'
    })

    client.post('/login', data={
        'username': 'testuser',
        'password': 'password123'
    })

    # Set the phone in session to avoid BuildError
    with client.session_transaction() as sess:
        sess['phone'] = '1234567890'  # Simulate phone in session

    # Log out and check for redirect
    response = client.post('/logout', follow_redirects=False)

    # Check if response is a redirect (302 status code)
    assert response.status_code == 302

def test_chat(client):
    """Test the chat functionality for a landlord."""
    # Register a landlord user
    client.post('/register', data={
        'phone': '1234567890',
        'username': 'landlorduser',
        'email': 'landlord@example.com',
        'password': 'password123',
        'user_type': 'landlord'
    })

    # Register a tenant user (for message recipient)
    client.post('/register', data={
        'phone': '0987654321',
        'username': 'tenantuser',
        'email': 'tenant@example.com',
        'password': 'password123',
        'user_type': 'tenant'
    })

    # Login as landlord
    client.post('/login', data={
        'username': 'landlorduser',
        'password': 'password123'
    })

    # Set up session to simulate being a landlord with all required session details
    with client.session_transaction() as sess:
        sess['username'] = 'landlorduser'
        sess['user_type'] = 'landlord'
        sess['phone'] = '1234567890'

    # Send a message as the landlord to a specific tenant
    response = client.post('/send_message/1', data={
        'message': 'Hello, tenant!',
        'pdf': (BytesIO(b'my file contents'), 'test.pdf')  # Simulate a file upload if required
    }, content_type='multipart/form-data', follow_redirects=False)

    # Check if the response is a redirect (302) indicating success or permission check passed
    assert response.status_code == 302 or response.status_code == 200
